+++ mooda/app/mooda_gui/__main__.py
@@ -4,7 +4,6 @@ from mooda.app.mooda_gui.widgets import MOODA
 
 
 def main():
-    print("Opening GUI. Please wait.")
     app = QApplication(sys.argv)
     md = MOODA()
     md.showMaximized()