+++ other_codes/gantt_chart_csv.py
@@ -1,5 +1,5 @@
 """
-It creates a plotly grantt graph from data of NetCDF files.
+It creates a plotly grantt graph from data of CSV files.
 The graph shows the intervals of time where there are data.
 With the grantt graph users can see where are the gaps.
 """
@@ -7,22 +7,24 @@ import glob
 import plotly
 import plotly.figure_factory as ff
 from mooda import WaterFrame
+from mooda.access import EGIM
 
 # Path of files
 # Where the files are:
-path_location = r"C:\Users\rbard\Desktop\Azores EGIM Data"
+path_location = r""
 # Where to save the result:
-save_location = r"C:\Users\rbard\Google Drive\Work\EmsoDev\server\www\html\graphs\azores"
+save_location = r""
 
 path_files = glob.glob(path_location)
-
 df = []
 for path_file in path_files:
 
+    if path_file[-1] != "v":
+        continue
+
     # Load data into a WaterFrame
     print("Loading data")
-    wf = WaterFrame()
-    wf.from_netcdf(path_file)
+    wf = EGIM.from_raw_csv("EMSO-Azores", path_file)
     print("Done")
     print(wf)
 
@@ -36,19 +38,14 @@ for path_file in path_files:
 
     intervals = wf.datetime_intervals(parameter)
 
-    if "R_4381-606" in path_file:
-        instrument = "Turbidimeter"
-    elif "R_37-14998" in path_file:
-        instrument = "CTD"
-    elif "NTURTD-64" in path_file:
-        instrument = "Oximeter"
-    elif "SBE54-0049" in path_file:
-        instrument = "Tsunameter"
+    print(path_file)
+    instrument = path_file.split(".")[-2].split("\\")[-1]
 
     for start, end in intervals:
         df.append(dict(Task=instrument, Start=start, Finish=end,
                        Resource=instrument))
 
-fig = ff.create_gantt(df, group_tasks=True, title="Data intervals", width=780,
+fig = ff.create_gantt(df, group_tasks=True, title="Data intervals", width=700,
                       height=380)
+fig['layout'].update(autosize=False, margin=dict(l=110))
 plotly.offline.plot(fig, filename=save_location)