+++ other_codes/plotly_TEMP.py
@@ -5,28 +5,38 @@ different instrument of the EGIM deployed in OBSEA.
 import plotly
 import plotly.graph_objs as go
 from mooda import WaterFrame
+from mooda.access import EGIM
 
 # The files that contain TEMP are from CTD, Tsunameter and Oxymeter
-path_location = r"WRITE YOUR PATH HERE"
-path_ctd = path_location + r"\OS_OBSEA_2016120120170426_R_37-14998.nc"
-path_oxymeter = path_location + r"\OS_OBSEA_2016120120170426_R_4381-606.nc"
+path_location = r""
+path_ctd = path_location + r"\CTD.csv"
+path_oximeter = path_location + r"\Oximeter.csv"
+
+output_file = r""
 
 # Load data into a WaterFrame
 print("Loading data from CTD")
-wf_ctd = WaterFrame()
-wf_ctd.from_netcdf(path_ctd)
+wf_ctd = EGIM.from_raw_csv("EMSO-Azores", path_ctd)
 print("Done")
-print("Loading data from the Oxymeter")
-wf_oxymeter = WaterFrame()
-wf_oxymeter.from_netcdf(path_oxymeter)
+print("Loading data from the Oximeter")
+wf_oxymeter = EGIM.from_raw_csv("EMSO-Azores", path_oximeter)
 print("Done")
 
 # Creation of a third Waterframe
+print("Creation of an other WaterFrame")
 wf_anal = WaterFrame()
+print("Concat CTD")
 wf_anal.concat(wf_ctd)
+print("Done")
+print("Rename parameters")
 wf_anal.rename("TEMP", "CTD")
+print("Done")
+print("Concat Oximeter")
 wf_anal.concat(wf_oxymeter)
+print("Done")
+print("Rename parameters")
 wf_anal.rename("TEMP", "OXYMETER")
+print("Done")
 
 # Calculation of correlation between the TEMP parameters
 print(wf_anal.corr("CTD", "OXYMETER"))
@@ -42,6 +52,10 @@ print("Done")
 print("Resampling data from Oximenter")
 wf_oxymeter.resample("H")
 print("Done")
+
+# Slice
+wf_ctd.slice_time(start="20170720000000", end="20180809000000")
+wf_oxymeter.slice_time(start="20170720000000", end="20180809000000")
 
 # Creation of traces for the plotly graph
 data_ctd = go.Scatter(x=wf_ctd.data.index, y=wf_ctd.data['TEMP'],
@@ -60,4 +74,4 @@ plotly.offline.plot(
     {
      "data": data,
      "layout": layout,
-    }, auto_open=True, filename=r"WRITE YOUR PATH\TEMP-OBSEA.html")
+    }, auto_open=True, filename=output_file)